requirements.txt
asgiref==3.2.3
blessings==1.7
bpython==0.18
certifi==2019.11.28
chardet==3.0.4
curtsies==0.3.1
dj-database-url==0.5.0
Django==3.0.2
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
greenlet==0.4.15
gunicorn==20.0.4
idna==2.8
psycopg2==2.8.4
psycopg2-binary==2.8.4
Pygments==2.5.2
PyJWT==1.7.1
python-decouple==3.3
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
wcwidth==0.1.8
whitenoise==5.0.1
